+++ HISTORY.rst
@@ -2,6 +2,9 @@
 History
 =======
 
+* Make ``MySQLCache.touch()`` return ``True`` if the key was touched, ``False`` otherwise.
+  This return value was missing since the method was added for Django 2.1.
+
 4.7.1 (2022-08-11)
 ------------------