+++ HISTORY.rst
@@ -2,6 +2,8 @@
 History
 =======
 
+* Add support for ``GroupConcat`` filters.
+
 * Make ``MySQLCache.touch()`` return ``True`` if the key was touched, ``False`` otherwise.
   This return value was missing since the method was added for Django 2.1.